pyproject.toml
[project]
name = "nl2sql-agent"
version = "0.1.0"
description = "NL2SQL sub-agent for Mako Group trading data"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.0.0",
    "google-cloud-bigquery>=3.25.0",
    "google-cloud-aiplatform>=1.60.0",
    "litellm>=1.40.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.0.0",
    "pandas>=2.0.0",
    "db-dtypes>=1.0.0",
    "mcp[cli]>=1.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.9.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["nl2sql_agent"]
namespaces = false

[tool.pytest.ini_options]
markers = [
    "integration: requires live services (LiteLLM proxy, BigQuery)",
]
addopts = "-m 'not integration'"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 88
src = ["nl2sql_agent", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "UP",   # pyupgrade (modernize syntax)
    "B",    # flake8-bugbear (common bugs)
    "I",    # isort (import sorting)
    "N",    # pep8-naming
    "S",    # flake8-bandit (security)
    "T20",  # flake8-print (no bare print)
    "PT",   # flake8-pytest-style
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific rules
]
ignore = [
    "E501",   # line length (formatter handles this)
    "S608",   # hardcoded SQL — expected for SQL generation tools
    "PT018",  # composite assertion — acceptable in our test style
    "PT006",  # parametrize names wrong type — style preference
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]  # allow assert in tests
"scripts/**/*.py" = ["T201"]  # allow print in scripts
"eval/**/*.py" = ["T201"]  # allow print in eval runner
"setup/**/*.py" = ["T201"]  # allow print in setup scripts

[tool.ruff.lint.isort]
known-first-party = ["nl2sql_agent"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
check_untyped_defs = true
disallow_untyped_defs = false
no_implicit_optional = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
check_untyped_defs = false

[[tool.mypy.overrides]]
module = [
    "google.adk.*",
    "google.cloud.*",
    "litellm.*",
    "structlog.*",
]
ignore_missing_imports = true

[tool.coverage.run]
source = ["nl2sql_agent"]
branch = true

[tool.coverage.report]
show_missing = true
fail_under = 80
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
]